refactor(fazbot.bot): refactor invoke to view
- Rename invoke modules to view
- Subclass nextcord.ui.View on BaseView
