+++ fazbot/bot/cog/help.py
@@ -4,7 +4,7 @@ import nextcord
 
 from fazbot.bot.cog._cog_base import CogBase
 from fazbot.bot.errors import ApplicationException
-from fazbot.bot.invoke.invoke_help import InvokeHelp
+from fazbot.bot.view.help import InvokeHelp
 
 
 class Help(CogBase):
