+++ fazbot/bot/view/_asset.py
