+++ fazbot/bot/view/_base_view.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, override
 
-from fazbot.bot.invoke._asset import Asset
+from nextcord.ui import View
+
+from fazbot.bot.view._asset import Asset
 
 if TYPE_CHECKING:
     from nextcord import Embed, File, Interaction
@@ -11,14 +13,29 @@ if TYPE_CHECKING:
     from fazbot.bot.bot import Bot
 
 
-class Invoke(ABC):
+class BaseView(View, ABC):
 
-    def __init__(self, bot: Bot, interaction: Interaction[Any]) -> None:
+    def __init__(
+        self,
+        bot: Bot,
+        interaction: Interaction[Any],
+        *,
+        timeout: float | None = 180.0,
+        auto_defer: bool = True,
+        prevent_update: bool = True,
+    ) -> None:
+        super().__init__(
+            timeout=timeout, auto_defer=auto_defer, prevent_update=prevent_update
+        )
         self._bot = bot
         self._interaction = interaction
 
     @abstractmethod
     async def run(self): ...
+
+    @override
+    async def on_timeout(self) -> None:
+        await self._interaction.edit_original_message(view=View(timeout=1))
 
     @staticmethod
     def _set_embed_thumbnail_with_asset(embed: Embed, filename: str) -> None:
