+++ tests/fazbot/bot/view/test_ingredient_probability.py
@@ -1,24 +1,23 @@
 from decimal import Decimal
 from typing import override
-from unittest import TestCase
+from unittest import IsolatedAsyncioTestCase
 from unittest.mock import AsyncMock, MagicMock, patch
 
-from fazbot.bot.invoke.invoke_ingredient_probability import InvokeIngredientProbability
+from fazbot.bot.view.ingredient_probability_view import IngredientProbabilityView
 
 
-class TestIngredientProbability(TestCase):
+class TestIngredientProbability(IsolatedAsyncioTestCase):
 
     @override
-    def setUp(self) -> None:
+    async def asyncSetUp(self) -> None:
         self.interaction = AsyncMock()
         self.asset = MagicMock()
         with patch("fazbot.bot.bot.Bot", autospec=True) as mock_bot:
             self.mock_bot = mock_bot
-        self.obj = InvokeIngredientProbability(
+        self.obj = IngredientProbabilityView(
             self.mock_bot, self.interaction, "1/1000", 500, 100
         )
         self.obj.set_assets(self.asset)
-        return super().setUp()
 
     # async def test_run(self) -> None:
     #     await self.obj.run()
